refactor: move main screen to main module and update its navigation
